+++ programs/incept/src/instructions/liquidate_comet.rs
@@ -571,7 +571,7 @@ pub fn execute(
         iasset_delta,
     });
 
-    let pool = token_data.pools[pool_index as usize];
+    let pool = token_data.pools[pool_index];
     let mut oracle_price = pool.asset_info.price.to_decimal();
     oracle_price.rescale(DEVNET_TOKEN_SCALE);
